Fix clang-tidy hits

+++ src/wui/quicknavigation.cc
@@ -259,10 +259,10 @@ void QuickNavigationWindow::rebuild() {
 			show_watch_window(dynamic_cast<InteractiveGameBase&>(ibase_),
 			                  MapviewPixelFunctions::calc_node_and_triangle(
 			                     ibase_.egbase().map(),
-			                     // Technically, a landmark is the top-left corner of the screen, but
-			                     // we like to pretend it's the screen center instead.
-			                     q.landmarks()[i].view.viewpoint.x + g_gr->get_xres() / 2,
-			                     q.landmarks()[i].view.viewpoint.y + g_gr->get_yres() / 2)
+			                     // Technically, a landmark is the top-left corner of the screen,
+			                     // but we like to pretend it's the screen center instead.
+			                     static_cast<int>(q.landmarks()[i].view.viewpoint.x + g_gr->get_xres() / 2),
+			                     static_cast<int>(q.landmarks()[i].view.viewpoint.y + g_gr->get_yres() / 2))
 			                     .node);
 		});
 		box.add(b);
